Add native signal delivery and a first-class process provider value

+++ rust/tests/node/process_tests.rs
@@ -92,12 +92,12 @@ fn process_identity_queries_use_platform_values() {
 fn process_kill_exposes_signal_zero_probe() {
     #[cfg(unix)]
     {
-        assert!(process::kill(process::pid(), Some(0)).unwrap());
-        assert!(process::kill(u32::MAX, Some(0)).is_err());
+        assert!(process::kill(process::pid() as f64, Some(0)).unwrap());
+        assert!(process::kill(u32::MAX as f64, Some(0)).is_err());
     }
     #[cfg(not(unix))]
     {
-        assert!(process::kill(process::pid(), Some(0)).is_err());
+        assert!(process::kill(process::pid() as f64, Some(0)).is_err());
     }
 }
